Allow JGit to perform Submodule operations.
The code in CliGitAPIImpl is sufficiently general as to not require
anything special from the command-line.  So, we can just move it to
LegacyCompatibleGitAPIImpl to allow to use JGit with submodules.
