+++ src/main/java/org/jenkinsci/plugins/gitclient/LegacyCompatibleGitAPIImpl.java
@@ -3,6 +3,7 @@ package org.jenkinsci.plugins.gitclient;
 import static java.util.Arrays.copyOfRange;
 import static org.apache.commons.lang.StringUtils.join;
 import hudson.model.TaskListener;
+import hudson.plugins.git.Branch;
 import hudson.plugins.git.GitException;
 import hudson.plugins.git.IGitAPI;
 import hudson.plugins.git.IndexEntry;
@@ -19,6 +20,7 @@ import org.eclipse.jgit.transport.URIish;
 
 import java.io.File;
 import java.io.IOException;
+import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.List;
@@ -71,6 +73,166 @@ abstract class LegacyCompatibleGitAPIImpl extends AbstractGitAPIImpl implements 
             throw new GitException("Couldn't check for .gitmodules", e);
         }
 
+    }
+
+    /**
+     * Get the default remote.
+     *
+     * @throws hudson.plugins.git.GitException if executing the git command fails
+     * @throws java.lang.InterruptedException if interrupted
+     * @return default git remote for this repository (often "origin")
+     */
+    public String getDefaultRemote() throws GitException, InterruptedException {
+        return getDefaultRemote("origin");
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * Set up submodule URLs so that they correspond to the remote pertaining to
+     * the revision that has been checked out.
+     */
+    public void setupSubmoduleUrls( Revision rev, TaskListener listener ) throws GitException, InterruptedException {
+        String remote = null;
+
+        // try to locate the remote repository from where this commit came from
+        // (by using the heuristics that the branch name, if available, contains the remote name)
+        // if we can figure out the remote, the other setupSubmoduleUrls method
+        // look at its URL, and if it's a non-bare repository, we attempt to retrieve modules
+        // from this checked out copy.
+        //
+        // the idea is that you have something like tree-structured repositories: at the root you have corporate central repositories that you
+        // ultimately push to, which all .gitmodules point to, then you have intermediate team local repository,
+        // which is assumed to be a non-bare repository (say, a checked out copy on a shared server accessed via SSH)
+        //
+        // the abovementioned behaviour of the Git plugin makes it pick up submodules from this team local repository,
+        // not the corporate central.
+        //
+        // (Kohsuke: I have a bit of hesitation/doubt about such a behaviour change triggered by seemingly indirect
+        // evidence of whether the upstream is bare or not (not to mention the fact that you can't reliably
+        // figure out if the repository is bare or not just from the URL), but that's what apparently has been implemented
+        // and we care about the backward compatibility.)
+        //
+        // note that "figuring out which remote repository the commit came from" isn't a well-defined
+        // question, and this is really a heuristics. The user might be telling us to build a specific SHA1.
+        // or maybe someone pushed directly to the workspace and so it may not correspond to any remote branch.
+        // so if we fail to figure this out, we back out and avoid being too clever. See JENKINS-10060 as an example
+        // of where our trying to be too clever here is breaking stuff for people.
+        for (Branch br : rev.getBranches()) {
+            String b = br.getName();
+            if (b != null) {
+                int slash = b.indexOf('/');
+
+                if ( slash != -1 )
+                    remote = getDefaultRemote( b.substring(0,slash) );
+            }
+
+            if (remote!=null)   break;
+        }
+
+        if (remote==null)
+            remote = getDefaultRemote();
+
+        if (remote!=null)
+            setupSubmoduleUrls( remote, listener );
+    }
+
+    private String pathJoin( String a, String b ) {
+        return new File(a, b).toString();
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * Fixes urls for submodule as stored in .git/config and
+     * $SUBMODULE/.git/config for when the remote repo is NOT a bare repository.
+     * It is only really possible to detect whether a repository is bare if we
+     * have local access to the repository.  If the repository is remote, we
+     * therefore must default to believing that it is either bare or NON-bare.
+     * The defaults are according to the ending of the super-project
+     * remote.origin.url:
+     *  - Ends with "/.git":  default is NON-bare
+     *  -         otherwise:  default is bare
+     *  .
+     */
+    public void fixSubmoduleUrls( String remote,
+                                  TaskListener listener ) throws GitException, InterruptedException {
+        boolean is_bare = true;
+
+        URI origin;
+        try {
+            String url = getRemoteUrl(remote);
+            if (url == null)
+                throw new GitException("remote." + remote + ".url not defined in workspace");
+
+            // ensure that any /.git ending is removed
+            String gitEnd = pathJoin("", ".git");
+            if ( url.endsWith( gitEnd ) ) {
+                url = url.substring(0, url.length() - gitEnd.length() );
+                // change the default detection value to NON-bare
+                is_bare = false;
+            }
+
+            origin = new URI( url );
+        } catch (URISyntaxException e) {
+            // Sometimes the URI is of a form that we can't parse; like
+            //   user@git.somehost.com:repository
+            // In these cases, origin is null and it's best to just exit early.
+            return;
+        } catch (Exception e) {
+            throw new GitException("Could not determine remote." + remote + ".url", e);
+        }
+
+        if ( origin.getScheme() == null ||
+             ( "file".equalsIgnoreCase( origin.getScheme() ) &&
+               ( origin.getHost() == null || "".equals( origin.getHost() ) )
+             )
+        ) {
+            // The uri is a local path, so we will test to see if it is a bare
+            // repository...
+            List<String> paths = new ArrayList<String>();
+            paths.add( origin.getPath() );
+            paths.add( pathJoin( origin.getPath(), ".git" ) );
+
+            for ( String path : paths ) {
+                try {
+                    is_bare = isBareRepository(path);
+                    break;// we can break already if we don't have an exception
+                } catch (GitException e) { }
+            }
+        }
+
+        if ( ! is_bare ) {
+            try {
+                List<IndexEntry> submodules = getSubmodules("HEAD");
+
+                for (IndexEntry submodule : submodules) {
+                    // First fix the URL to the submodule inside the super-project
+                    String sUrl = pathJoin( origin.getPath(), submodule.getFile() );
+                    setSubmoduleUrl( submodule.getFile(), sUrl );
+
+                    // Second, if the submodule already has been cloned, fix its own
+                    // url...
+                    String subGitDir = pathJoin( submodule.getFile(), ".git" );
+
+                    /* it is possible that the submodule does not exist yet
+                     * since we wait until after checkout to do 'submodule
+                     * udpate' */
+                    if ( hasGitRepo( subGitDir ) ) {
+                        if (! "".equals( getRemoteUrl("origin", subGitDir) )) {
+                            setRemoteUrl("origin", sUrl, subGitDir);
+                        }
+                    }
+                }
+            } catch (GitException e) {
+                // this can fail for example HEAD doesn't exist yet
+            }
+        } else {
+            // we've made a reasonable attempt to detect whether the origin is
+            // non-bare, so we'll just assume it is bare from here on out and
+            // thus the URLs are correct as given by (which is default behavior)
+            //    git config --get submodule.NAME.url
+        }
     }
 
     /** {@inheritDoc} */
